remove astype(int64)

+++ python/sgl_jax/srt/layers/attention/flashattention_backend.py
@@ -158,13 +158,13 @@ class FlashAttention(AttentionBackend):
                 end = (i + 1) * per_dp_loc_len
                 rank_cache_loc = batch.cache_loc[start:end]
                 mapping = swa_mapping[i] if isinstance(swa_mapping, list) else swa_mapping
-                rank_swa_cache_loc = mapping[rank_cache_loc].astype(np.int64)
+                rank_swa_cache_loc = mapping[rank_cache_loc].astype(np.int32)
 
                 remainder = len(rank_swa_cache_loc) % self.page_size
                 if remainder > 0:
                     pad_len = self.page_size - remainder
                     rank_swa_cache_loc = np.concatenate(
-                        [rank_swa_cache_loc, np.zeros(pad_len, dtype=np.int64)]
+                        [rank_swa_cache_loc, np.zeros(pad_len, dtype=np.int32)]
                     )
 
                 rank_swa_selected = rank_swa_cache_loc[:: self.page_size]
